Refactor wine display components and improve expert service
- Introduced a new `WineCardComponent` to encapsulate wine card logic and styling, replacing inline code in `dashboard` and `wine-list` components.
- Updated `dashboard.component.html` and `wine-list.component.html` to use the new `app-wine-card` component for better maintainability.
- Removed redundant type color logic from `dashboard.component.ts` and `wine-list.component.ts`, moving it to `wine-card.component.ts`.
- Enhanced `ExpertService` to handle SSE responses more effectively, including status updates and error handling.
- Improved UI text for clarity and consistency in the `expert.component.html`.
- Added Vinmonopolet search functionality to wine cards, allowing users to search directly from the wine card.

+++ api/Models/ExpertModels.cs
@@ -7,7 +7,7 @@ public record ExpertSessionSummary(
     string?  Title,
     DateTime CreatedAt,
     DateTime UpdatedAt,
-    int      MessageCount
+    long     MessageCount
 );
 
 // ── Full session with messages ──────────────────────────────────────────────
